pyproject.toml
[project]
name = "stock-analysis"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.96.0",
    "openai>=2.32.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "xlrd>=2.0.2",
]
